Remove .gemrc
One of the Ruby committers gave a talk at RubyConf 2018 and talked about the improvements they're making to the Ruby documentation system. He encouraged everyone to turn docs back on for gems, so I'm removing the no-doc settings from my .gemrc. That's the only thing in it though, so I'm just deleting the file entirely.