Get GPG fully working
I used GitKraken to set up GPG signing for my Git commits. It always works in GitKraken and works most of the time using Git on the command line, but there were cases where it failed with the following error:
```
error: gpg failed to sign the data
fatal: failed to write commit object
```

This commit includes some additional configuration that fixes the above error and gets GPG signing working fully for the command line.

I followed the accepted answer on this StackOverflow post: https://stackoverflow.com/questions/39494631/gpg-failed-to-sign-the-data-fatal-failed-to-write-commit-object-git-2-10-0#answer-40066889