Symlink Development directory in iCloud to the home dir
I've hesitated to put my Development directory in iCloud in the past because `cd`ing to the iCloud directory is a pain. I'd like all of those files backed up to the cloud though. I realized moving all of my development stuff to the iCloud directory and then symlinking them into the home directory is a good solution.