Update Rubocop config
Disable Metrics/BlockLength for the `spec` and `config` directories. Both of these directories contain files that can have really long blocks. In those contexts though, it's not bad to have long blocks.

Increase the maximum for RSpec/ExampleLength from 5 to 10. I agree with keeping RSpec examples short, but 5 seems to short to enforce with Rubocop.