Remove auto-pairs plugin from nvim
There were times when it was helpful to automatically have closing parenthesis, quotes, etc. added for me, but there were a lot of other times where it just got in the way because I only wanted to add one.

Overall, I don't think it was making my editing experience better.