Fix behavior of swap-window keybinding
Previously, the `swap-window` command would swap the windows you told it to and keep your view on the original window you were looking at. They changed this in recent tmux versions where `swap-window` switches the windows and keeps your view on the same slot instead of the same window. So when you swap windows, your view gets switched to the other window. I find this unintuitive and not what I want. I want to stay on the window I'm currently looking. Adding the `previous-window` and `next-window` commands to the keybindings restores the old behavior.

[Credit to Steven Lu for this idea.](https://unix.stackexchange.com/a/525770/392073)