Don't automatically update ctags on Nvim startup
I originally added this command because it was a convenient way to keep tags updated for projects I was working on. The problem is that updating tags can take several seconds on larger projects. This made starting up Neovim painfully slow on some projects and was interfering with my workflow. I'll either need to find a better way to automatically update tags or just do it manually. I don't use tags all that often, so, for now, I'm comfortable just updating them manually.