Remove tmux-resurrect
I didn't ever use it (I depend on tmuxinator instead), and it seemed 
like it was causing issues. I saw error messages from it somewhat regulary,
and the tmux server starting crashing occasionally. I don't know that the
server crashes are the fault of tmux-resurrect, but I'm suspicious that it
might be.