Remove Brave browser from Mac Setup
I seldom ever use it. In theory, it's nice to have a version of chromium
that respects my privacy. In practice, I don't ever use it because Firefox
works very well the vast majority of the time. When I need to view something
with Chrome, like I'm testing an application I'm developing, I think it
makes the most sense to use actual Chrome.

I don't mind using Chrome on a limited basis.