Use the built in Mac screenshot app instead of Snip
The built in screenshot app does everything Snip does plus a little more.
There are some things about it that don't seem as convenient (for example,
it doesn't automatically copy images to the clipboard), but overall it
seems like it might be better to just use the built in Mac utility rather
than installing a new application.