Remove Lacona from setup
Lacona was nice in a lot of ways, but the new version lacks many of the
features that made it meaningfully more useful than Spotlight. I tried
switching back to the previous version, but it didn't work.

I've decided to go without Lacona for a while and see how well Spotlight
works. That will require a couple of small changes to my workflow, but
nothing huge, and I think it will be worth it to eliminate a dependency
and use a default MacOS tool instead.