Revert "Turn on neovim spellcheck"
This reverts commit fc80e0ac72962c1b30a23cf9066a1e3f9022c24c.

Turning this setting on makes spellcheck too aggressive for my taste.
It was marking words as misspelled in contexts where it should have
realized they were code and weren't misspelled. There were some
contexts where it understood the difference between prose and
code, but it got it wrong often enough, that I don't think it's worth
having on. I still get spellchecking in some contexts even with it off.