Revert "Turn on neovim spellcheck"
This reverts commit c5997af9740ab8383a33bb6bb5f6906e1dce1797.

Turning this setting on makes spellcheck too aggressive for my taste.
It was marking words as misspelled in contexts where it should have
realized they were code and weren't misspelled. There were some
contexts where it understood the difference between prose and
code, but it got it wrong often enough, that I don't think it's worth
having on. I still get spellchecking in some contexts even with it off.