Remove Dropbox from Mac setup
We use Dropbox to share certain files for setting up local projects and
we have scripts that reference those files in the `~/Dropbox` folder, so
that's why I originally installed it. I don't actually like having the client
installed though, so I've decided I'll just manually download the aj-dev
directory from Dropbox, and if I need to, I can occasionally download it
again to keep it up to date since Dropbox won't be installed to do it
automatically now.