Add ftplugin for nerdtree so it isn't expanded by golden-ratio plugin
While this prevents golden-ratio from expanding the nerdtree pane
directly, the pane still gets indirectly expanded when other windows
are resized to 2/3. When that happens, the nerdtree window expands
to 1/3. Ideally, I'd like it to stay the same size as when it originally opens
(which seems to be about 1/8 of the screen), but I'm not sure there's a
way to do that. See issue #3 [1] for more details.

[1]: https://github.com/roman/golden-ratio/issues/3