Use a Chrome extension for markdown previews instead of nvim plugin
I never succeeded in finding a markdown preview plugin that I really
liked and worked well. The one I'm removing in this commit works fine
when it works, but it crashes constantly and I have to keep reopening
the preview.

I'm planning on using a Chrome extension that I really like instead.