Use the YouTube Music website instead of a desktop player
I was using the Google Play Music Desktop Player [1] for a while and it
was mostly good, but I ran into issues regularly. So I tried using the
YouTube Music Desktop App [2] instead, but it also had issues. So I'm
trying out just using the official web app instead. I'm going to run it
using Brave Browser as the dedicated browser for it so that I can
map a keybinding to it.

[1]: https://github.com/MarshallOfSound/Google-Play-Music-Desktop-Player-UNOFFICIAL-
[2]: https://ytmdesktop.app/