Add Jabra Direct to Mac setup
I've decided I prefer using my Jabra headset through the USB link instead
of through a generic bluetooth connection. It makes it easier to tell
how much battery life is left, and it prevents the volume input level
from changing every time I turn the headset on and off.

The downside is I have to manually switch the audio input and output
in the MacOS settings, but it's pretty easy to do that, and I'd rather do
that than have to check the input volume every time.