Update tmux config to put rails server pane and terminal pane together
The 2 panes I use the most (outside of my text editor) are the Rails
server and the terminal. I separated them because it made more sense
to me to put all the servers together and put the Rails console with
the terminal. That setup might make more sense categorically, but
practically, I reach for the terminal pane and the Rails server pane more
than anything else, and I often find myself moving each pane's respective
window to the second position based on what I'm doing at the time.
By putting both the terminal and the Rails server in the second window,
I'll have the easiest access to the panes I need and it will reduce
development friction by a little bit.