Remove some outdate Mac setup instructions
I don't use a 3rd party screenshot application anymore. I've discovered
(via Brandon's recommendation) that the built in Mac screenshot app
works great. So I'd rather not pull in a 3rd party app.

I used Vimac for a bit, and it was a cool idea, but I don't think it worked
reliably enough to be worth adding to my workflow.