Manually set Node version for coc.nvim
The current version of coc.nvim requires node >= v12.12.0. In projects
where we have the Node version set to an older version, coc.nvim didn't
work because it needed a newer node version. By manually setting
the path, I can use a more recent version of Node installed via Brew
to satisfy the version requirement.

The downside to this approach is that it's hardcoded to a certain version.
When I update Node with `brew upgrade`, this setting will have to be
updated. I'd like to find a better solution at some point, but this is
acceptable to me for now since I don't run `brew upgrade` all that
often.