Display column numbers in Vim status line
I realized I do use the column number occasionally. The instance I ran
into was wanting to know how long the line I was on was. I think of that
when a line starts to look like it might be too long. Maybe I should just
go off of how long the the line feels, but it is sometimes useful to know
exactly how many characters are in the line.

In the past, I would just `$` to the end of the line and then check the
column number in the status line. It seemed like it would be helpful
to know how long a line was no matter where on the line my cursor
was so I added both the column of the cursor and the total length of
the line. It makes that section of the status line a little longer than I'd
like, but I think the extra info will be worth it. We'll see.

I put the column first because it's quicker to go to the edge of the screen
with my eyes, and I'm sure I'll check the line number info more often
than I'll check the column number info.