Remove old Git pre-commit hook
In theory, checking for style violations and failing tests on each commit
sounds nice, but it's just too slow. Especially on larger projects. Additionally,
there are times where I intentionally want to commit even though I
have style violations and failing tests. (One example would be after
running a generator and I want to commit the changes exactly as they
were generated and clean them up in a separate commit so it's clear
which code was generated and which was mine. These will often have
style violations that I don't want to fix in the first commit.)

It's for these reasons that I stopped using this pre-commit hook a while
ago, and I don't foresee myself ever going back. So it makes sense to
just delete it at this point. Even if I do want to start doing this again, it
wouldn't be that hard to recreate it.