Remove old Ruby config for asdf
I don't remember why I added this. I don't see anything about it in the
asdf or asdf-ruby docs. I recently had some trouble with finding a usable
version of openssl when I was installing Erlang; I wonder if this line was
the reason why. I'm going to remove it and see if anything breaks.