Remove gpg program path from global Git config
I don't see a reason to manually specify the path to the gpg program.
If you don't specify the path, it will just use whatever gpg it can find in
the $PATH. That seems like the behavior I'd prefer.

Additionally, this gpg path was outdated. The correct path is
`/opt/homebrew/bin/gpg`.