Remove some MacOS settings automation
While I like having these automated, I feel hesitant about the reliability
of these commands. Some of the ones I removed are either no longer
relevant or don't work for some reason. For example, this commit [1]
shows how the command to show the Library folder (the one I was using)
was broken in Catalina.

I kept some of the commands around because there's no way to change
the settings through the GUI that I'm aware of or the command is
particularly helpful (like setting the Dock to an exact size), but I think
I'm going to just make most of these changes manually going forward.
I might change my mind if I become convinced that these commands
are reliable enough. It would be nice to automate them all.

It also seemed like figuring out the correct commands might be too
time consuming.

This was the original article I used to start automating settings:
https://pawelgrzybek.com/change-macos-user-preferences-via-command-line/

This website seems like it might be pretty helpful if I do decide to
automate in the future: https://macos-defaults.com

[1]: https://github.com/mathiasbynens/dotfiles/commit/13e29f408fc360360b4468aef2c318dff287c7d2