Improve colors in neovim
This was recommended in the README for the rakr/vim-one theme I'm
using. Before turning it on, the background in neovim was a light gray
that made it harder to read. After turning `termguicolors` on, the background
turned to a white that matched the rest of the terminal and was easier
to read.