Revert "Add Fig to setup"
This reverts commit f30681652e1cf911526ec5dbfa21792778078347.

I'm removing Fig because it caused issues with vim-tmux-runner [1] [2] [3].
The founders of Fig are trying to work with Chris Toomey to find a
solution though [3], so I could get it working if I really wanted to I bet.

For now, I'm just going to remove Fig; vim and tmux navigation is
more important.

The thing that attracted me to Fig was the command line autocomplete, but
I'm not sure that's worth the cost. Fig does a lot more than I actually
want it to, and it seems to be a fairly complex tool. Dealing with it doesn't
seem worth the value it would create for me. I'll pass.

[1]: https://github.com/christoomey/vim-tmux-navigator/issues/339
[2]: https://github.com/christoomey/vim-tmux-navigator/issues/317
[3]: https://github.com/christoomey/vim-tmux-navigator/issues/295