Improve background color in LunarVim theme
Disabling the background with `disable.background = true` mostly worked;
most of the background areas of the editor inherited the color of the
terminal. The problem was that certain areas did not (such as Telescope
pickers). This wasn't a big deal but it didn't look great and it negatively
affected legibility in those areas (the reason I disabled the background
color in the first place). Changing the `bg` color to white makes it so
all of the background areas are white, fully fixing the problem.