Use bufferline plugin for tabline instead of build in buffer line
I could just configure a tabline that I like with herline, but heirline doesn't
come with much of anything built in; you have to build it all yourself.
It's a lot simpler for me to just use bufferline since I already know I like
it and I even tweaked it a little to match my taste.