Improve Telescope keybinding
I'm used to `<leader>ft` to search for text, but AstroNvim uses `fw`
(find words), I'm going to go ahead and switch to that so that I match
AstroNvim better. It won't take much effort to remap that in my brain
and the keystrokes are actually easier for my fingers. And then I don't
have to remap `<leader>fW` for things to stay consistent.