Disable Reek in null-ls
I've decided I don't want Reek feedback in my editor. It's helpful to
consider what Reek has to say, but the type of stuff Reek comments
on is too nuanced to have a linter error in the editor for it. I added a
todo item in my PR guide to run Reek. So basically Reek will be a
computer code reviewer, but I won't ask for it's feedback constantly
as I'm writing code.