Exclude VCR cassettes from primary file finder
I'm almost never interested in finding a cassette and they clutter up
searches. In fact, they usually appear at the top because they don't
have underscores in the filename, so they match a search of `planscontroller`
better than `plans_controller.rb` so I have to scroll through a few
cassette files before getting to the file I actually want.

In the event that I actually want to look for a given cassette, I added a
specialized finder for it.