Revert "Replace `heirline` winbar with `utilyre/barbecue.nvim`"
This reverts commit 2cbbcfd74ba34863203c41c5554de69d8e7c232a.

While using Neovim to rebase a branch, Barbecue threw an exception every
time I moved my cursor. I tried finding a similar issue in the repo on
GitHub but didn't see anything. I'm going to remove Barbecue and use
Heirline to achieve a similar winbar. The AstroNvim docs [1] have a 
recipe for it.

[1]: https://astronvim.com/Recipes/status#replicate-visual-studio-code-winbar