Revert "Replace `heirline` winbar with `utilyre/barbecue.nvim`"
This reverts commit bc443fa0220d6dfaec9db8827dcda854b3090783.

While using Neovim to rebase a branch, Barbecue threw an exception every
time I moved my cursor. I tried finding a similar issue in the repo on
GitHub but didn't see anything. I'm going to remove Barbecue and use
Heirline to achieve a similar winbar. The AstroNvim docs [1] have a 
recipe for it.

[1]: https://astronvim.com/Recipes/status#replicate-visual-studio-code-winbar