Update Alfred workflow that opens GitKraken
I couldn't commit using the GitKraken UI because Husky would fail. However,
if I launched GitKraken with a different default shell (like bash or zsh),
it worked fine. This updates the Alfred workflow so that it first launches
GitKraken with an environment variable specifying a different default
shell.

This issue [1] was helpful.

[1]: https://github.com/typicode/husky/issues/243