Remove Deseret Book BTT preset and rename personal preset
Duplicating the preset (which I did when I created the Deseret Book
preset was a bad idea. If I ever want to change a configuration value
for both my personal and work preset, I'll have to change it in 2 places.
The approach I'll use instead is to have a base preset and then I'll
create a Deseret Book preset that overrides a few of the bindings
and settings.