Remove Reek
I haven't had a good experience with Reek. Too many of it's recommendations
aren't helpful. It's not worth the cost of working with it. Rubocop, as a
counter example, makes mostly useful recommendations. It's uncommon
for me to disagree with a Rubocop recommendation, and Rubocop
recommendations are more fine grained so it's a lot easier to stomach
a recommendation I don't like. Reek tries to make recommendations
that are much higher level and require quite a bit more refactoring to
follow.