Remove Obsidian config files
I don't think it will work very well to add these to my dotfiles repo
and then symlink them because the symlinks need to be different on
different machines. I think I'll just leave them to be synced by iCloud.
I'll lose versioning, but that's ok. As long as I have the configs synced
and backed up, that's the important thing.