Replace Maccy with BTT clipboard manager
I tried this switch before, but didn't like BTT's clipboard manager as
much. However, it's gotten some improvements recently and the
developer says there's more planned for next year, so I'm going to
give it another try. Maccy is still being buggy for me.

One of the missing features in BTT that stopped me before was I
couldn't hit Cmd-1 or Cmd-2 to copy/paste an item in the list; that's
been added now though.