Switch to Microsoft Copilot desktop app instead of web app
I was using the web app because that's what allowed me to sign in with
my company account, but it was an inferior experience, and it didn't
work in Zen Browser for some reason. So I decided to use the desktop
app and just sign in with my personal account. That seems like the best
option.