Prepare for AstroNvim 5.0 upgrade
There were a number of plugin changes in version 5, so these
configuration values are no longer relevant.

The biggest change is switching from Telescope to snacks.nvim.
I had a lot of Telescope config that might need to be migrated
to snacks.nvim. It's possible that snacks.nvim does part of what
I manually configured Telescope to do.