Remove Cursor extensions.json file
I didn't seem to work. After adding the file and opening Cursor, it didn't
list the expected extensions or anything. I still had to manually install
them. Not a big deal to do when setting up the application for the first
time. And the configuration for the extensions is in settings.json.