Add Claude Code
My plan is to use Neovim as my primary editor. I don't want to switch
to a VS Code fork like Cursor. But I do want the agentic tooling that
applications like Cursor provide. My plan was to use Cursor as a
secondary editor when I wanted an AI agent, but I think I actually like
the user experience of Claude Code a bit better. Plus, it runs in the
terminal so it'll be right next to Neovim in tmux; I think that will feel
a lot smoother than switching between iTerm and Cursor. It might
end up being more expensive than Cursor, but I'll experiment and see.

I disabled tools in Avante, because I only plan to use Avante for small
edits and asking questions. If I want tooling (searching files, editing
mutliple files, etc.), I'll use Claude Code. I found the experience in
Avante for this kind of thing to be quite poor.

In all 3 of these tools (Cursor, Avante, Claude Code), the model being
used is the same (Claude Sonnet and Claude Haiku), so the quality
of the output should be very similar; it's primarily about what scenario
gives me the UX. For now, I think Neovim and Claude Code is the
best choice.