Replace `vim-rails` with `other.nvim`
The only thing I used in `vim-rails` was the related file command to
open a related test. It's a big plugin that adds a lot. I don't love including
it for only one feature. I saw `other.nvim` added to the community configs
and it looks a good replacement. It's only job is to open related files,
and it does it better than `vim-rails` in my opinion.