Add Notification and Stop hooks for Claude Code
I tried and failed to get Claude Code's built in notifications to work. So
I decided to use the `terminal-notifier` gem to manually send a
notification when Claude Code stops or hits a notification event.

The osascript part is so that it only sends the notification if iTerm isn't
the focused application. A little ugly, but it works.