Set D and U to scroll in Mouseless
I'm going to use this instead of Homerow. I like Homerow slightly
better, but it's a paid app and I don't use it that often. I think, for
how often I want to use the keyboard to scroll, Mouseless is an
adequate option. This will allow me to eliminate a paid dependency
and still have the features I want.