Migrate from setup script to setup instructions
The setup script file is nice because it's faster and automated, but I don't
think it will be that much faster, and I like the idea of being more aware
of everything I'm setting up. It will make it easier to customize for the
machine I'm setting up and it's a good review of everything I'm installing
every few years.