Revert "Add serena_config.yml"
This reverts commit 0f22737f984b9f22b9c049fc76e6fbf4ddf6ba53.

I originally added this because I configured the open dashboard
on launch configuration value, but the config file includes all of
the projects you've activated with Serena, so the file is constantly
dirty. I don't want to keep committing that. And the setting I changed
is a very minor one that's easy to remember to update on a new
machine.