Adjust hyper key modification to fix app switcher issue
I was having an issue with the app switcher. I would hit cmd+tab and
it would stay on the app switcher instead of switching to the next
application like normal. Sometimes it would switch after a couple of
seconds. Other times I had to hit a key to make it switch. Claude
recommended adjusting the `modifiers` attribute and it seems to
have resolved the issue.