Change back to native spotlight keybinding
I turned off the native macOS keybinding to open Spotlight
(cmd+space) and replaced it with a trigger in Better Touch Tool. This
was so I could map cmd+space to cmd+space in Parallels and VMWare
for Omarchy. Since I've decided it's not worth trying to use Omarchy in
a VM, I can remove this BTT configuration and use the macOS native
feature again.