Remove list of optional modifiers for hyper key modification
I started to have a similar problem in Zen as I was having with Spotlight
(described earlier in 37c1d0bb). Hitting ctrl-tab should cycle tabs but
it was getting stuck on the tab switcher and I had to hit another key
to get it to switch. Making this `modifiers` list in Karabiner empty fixed
the issue in Zen and everything else seems to still be working correctly.