Remove Rails specific agents
I'm just not sure if these agents are the most effective way to go about
this work. I pulled them from Obie Fernandez's Claude on Rails project,
but he doesn't use that anymore. And I'm not sure enough myself if it's
best to delegate this type of work to these types of agents. It feels best
to remove them.