Remove code-review plugin
I like some aspects of this plugin but it doesn't work super well
with my flow because you have to explicitly tell it to complete
the review if the PR is in draft (which mine always are when I
want Claude to review), and it wants to comment on the PR in
GitHub (which I don't want).

I'm going to use the `pr-review-toolkit:review-pr` command and my
custom code-reviewer agent instead.